+++ extensions/vscode/src/commands/remove-brick.ts
@@ -1,11 +1,11 @@
-import * as _ from "lodash";
+import { isNil } from "lodash";
 import * as vscode from "vscode";
 import * as path from "path";
 import { masonRemove } from "../mason";
 
 export const removeLocalBrick = async () => {
   const cwd = vscode.workspace.workspaceFolders?.[0].uri.fsPath;
-  if (_.isNil(cwd)) {
+  if (isNil(cwd)) {
     return;
   }
 
@@ -20,7 +20,7 @@ export const removeLocalBrick = async () => {
   }
 
   const brick = await promptForBrickName();
-  if (_.isNil(brick)) {
+  if (isNil(brick)) {
     return;
   }
 
@@ -29,12 +29,12 @@ export const removeLocalBrick = async () => {
 
 export const removeGlobalBrick = async () => {
   const cwd = vscode.workspace.workspaceFolders?.[0].uri.fsPath;
-  if (_.isNil(cwd)) {
+  if (isNil(cwd)) {
     return;
   }
 
   const brick = await promptForBrickName();
-  if (_.isNil(brick)) {
+  if (isNil(brick)) {
     return;
   }