+++ packages/mason/lib/src/generator.dart
@@ -15,6 +15,11 @@ import 'package:pool/pool.dart';
 
 part 'hooks.dart';
 
+// This is intentionally a getter instead of a constant since nested hook
+// execution depends on being able to detect the runtime dynamically.
+// ignore: prefer_const_constructors
+bool get _isAotCompiled => bool.fromEnvironment('dart.vm.product');
+
 final _descriptorPool = Pool(32);
 final _partialRegExp = RegExp(r'\{\{~\s(.+)\s\}\}');
 final _fileRegExp = RegExp(r'{{%\s?([a-zA-Z]+)\s?%}}');
@@ -639,10 +644,11 @@ extension on HookFile {
   }
 
   File module(String checksum) {
+    final extension = _isAotCompiled ? 'aot' : 'dill';
     return File(
       p.join(
         buildDirectory.path,
-        '${p.basenameWithoutExtension(path)}_$checksum.dill',
+        '${p.basenameWithoutExtension(path)}_$checksum.$extension',
       ),
     );
   }