+++ packages/mason/lib/src/hooks.dart
@@ -252,10 +252,11 @@ class GeneratorHooks {
 
     if (uri == null) throw HookMissingRunException(hook.path);
 
+    final mode = _isAotCompiled ? 'aot-snapshot' : 'kernel';
     final progress = logger?.progress('Compiling ${p.basename(hook.path)}');
     final result = await Process.run(
       'dart',
-      ['compile', 'kernel', uri.toFilePath()],
+      ['compile', mode, uri.toFilePath()],
       runInShell: true,
     );