+++ packages/mason/test/src/bundler_test.dart
@@ -116,7 +116,6 @@ void main() {
         expect(bundle.changelog, isNull);
         expect(bundle.license, isNull);
         expect(bundle.files.length, equals(1));
-        expect(bundle.hooks.length, equals(4));
         final expectedFiles = [
           'post_gen.dart',
           'pre_gen.dart',
@@ -127,6 +126,7 @@ void main() {
           final hookFile = bundle.hooks[i];
           expect(hookFile.path, equals(expectedFiles[i]));
         }
+        expect(bundle.hooks.length, equals(4));
       });
 
       test('returns a MasonBundle when brick exists (plugin)', () {